Add files via upload
Project Description: 
Budget Manager is a Console Based that is used to manage is budget and track the income 
and expenses. 
 
Project Objectives: 
• Create an User Account add User Details. 
• Enter Income or Expense Information 
• View Income or Expense or Net Balance Summary Information 
 
 
Project Scope: 
To track and manage the income and expense transaction of a person Since it is console 
based project and it can be run on any ide like eclipse ,vs code . 
1. User Profile Table Block 
a. User 
b. User Details 
2. Transaction Table Block 
a. Income Transaction 
b. Expense Transaction